requirements.txt
flask
python-dotenv
flask-cors
pymongo
azure-storage-blob
azure-cosmos
azure-identity